# improve: speed up from_df and from_json