Speeds up construction of `Annotation` instances when using the `from_df` and `from_json` methods by avoiding repeated calls to `SortedSet.__setitem__`. In practice, this leads to speeds up of on the order 2X, especially when dealing with large numbers of segments. E.g.

    >>> def using_setitem(segs): 
    >>>     sdict = SortedDict() 
    >>>     for seg in segs: 
    >>>         sdict[seg] = {} 
    >>>     return sdict
    >>> def using_constructor(segs):
    >>>     return SortedDict({seg : {} for seg in segs})
    >>> segs = [Segment(n, n+1) for n in range(0, 100, 2)]
    >>> %timeit using_setitem(segs)
    135 µs ± 2.06 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
    >>> %timeit using_constructor(segs)                                                                                        
    54.6 µs ± 216 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)

I have also verified this speedup using `from_df` and `from_json` on actual data (diarization and phone alignments).

A secondary benefit is the refactoring of code so that majority of the code now lives in a `from_records` method that is called by `from_df`/`from_json`.