@hbredin 2020-01-27
Merging, thanks! 

I'd love to hear more about your tests on actual data and how faster that is.
I guess we should use this in [`pyannote.database.utils.load_rttm`](https://github.com/pyannote/pyannote-database/blob/develop/pyannote/database/util.py#L429) as well.