@nryant 2020-01-27
@hbredin 

I've been playing around with using [airspeed velocity](https://github.com/airspeed-velocity/asv) for benchmarking and could submit a pull request adding this to the `pyannote.core` repo. As I'm still very much a novice with that frameworks, the benchmarks are quite simple and involve running `from_df ` and `from_json` each on the following:

- a phone level alignment selected from one 2 minute clip from the BU corpus (about 2,000 segments)
- fake diarization data involving 8 segments by 2 speakers

So, two extremes in terms of input sizes and label set sizes.