@hbredin 2020-01-28
Anything that would ultimately make the pyannote ecosystem faster is interesting. 

Speed benchmarks are obviously one of those 👍 
How would that work? Github Actions?