@nryant 2020-01-28
Both NumPy and SciPy use [airspeed velocity](https://github.com/airspeed-velocity/asv, which is how I came across it. For a given project, you create a `benchmarks/` directory at the top level of the repo with the following contents:

- asv.conf.json  --  a config file that specifies where to find the repo (URL or path), which branches to benchmark by default, which versions of Python to test on, whether to use virtualenv or conda, etc
- benchmarks/  --  an importable package containing the benchmarks to be run

For concrete examples, see:

- [numpy](https://github.com/numpy/numpy/tree/master/benchmarks)
- [scipy](https://github.com/scipy/scipy/tree/master/benchmarks)

As you can see from the examples above and from the [guidelines](https://asv.readthedocs.io/en/stable/writing_benchmarks.html), writing benchmarks is a lot like writing tests using `pytest`  --  just write a function or method whose name begins with ``time_`` and airspeed velocity will discover it and time it using the same heuristics as iPython's `%timeit` module. The benchmark suite will be run in a virtual environment `asv` sets up and the results stored in JSON files.  The results can then be viewed from the command line or via a simple website that `asv` generates and which will look something like [this]( https://pv.github.io/numpy-bench/).

Of course, there's a lot more: support for memory as well as timing benchmarks, the ability to parameterize individual benchmarks, automatic regression detection, a development mode, convenience functions for comparing versions with cProfile, etc. Most of which I haven't used yet.

If this seems like something that would be of interest, I could create a pull request that populates a stub `benchmarks/` directory for `pyannote.core` along with a couple of sample benchmarks and then this could be fleshed out over time. Then we could move additional discussion about how to proceed to that pull request.