@hbredin 2020-01-29
Looks very interesting (if fully automated and maintained in the long term).
Feel free to create that pull request 👍